requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
ecdsa==0.19.0
fastapi==0.115.5
fastapi-utils==0.8.0
h11==0.14.0
httptools==0.6.4
idna==3.10
mysql-connector-python==9.1.0
mysqlclient==2.2.6
passlib==1.7.4
psutil==5.9.8
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.0
pydantic_core==2.27.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
watchfiles==0.24.0
websockets==14.1
